Automatically add an LVM managed /data partition
A formula wrapper around https://github.com/mikepea/auto-resize-data-fs

The purpose of this is to ensure that, if we have extra disks assigned
(either via EBS on AWS or Independent Disks in vCD), that we automatically
manage them being in /data

10% space is reserved, so that LVM snapshots can be taken.

There's some gnarly logic around mounting the LV, which essentially will
only happen if /data is empty. In short, it works out of the box for new
deploys, but a simple migration will need to be done if /data is already
populated on the root disk.